InventoryApp - iOS/inventory_api/requirements.txt
Flask
pymysql
python-dotenv
argon2-cffi
Flask-JWT-Extended
